Consolidate shared persistence code
