--- ClashOps/ClashOps/Persistence.swift
@@ -1,56 +1,0 @@
-//
-//  Persistence.swift
-//  ClashOps
-//
-//  Created by Matthew Riley on 2025-08-21.
-//
-
-import CoreData
-
-struct PersistenceController {
-    static let shared = PersistenceController()
-
-    static var preview: PersistenceController = {
-        let result = PersistenceController(inMemory: true)
-        let viewContext = result.container.viewContext
-        for _ in 0..<10 {
-            let newItem = FavDeck(context: viewContext) // ✅ use a Core Data entity
-            newItem.name = "Preview Deck"
-            newItem.cards = "Card1,Card2,Card3"
-        }
-        do {
-            try viewContext.save()
-        } catch {
-            let nsError = error as NSError
-            fatalError("Unresolved error \(nsError), \(nsError.userInfo)")
-        }
-        return result
-    }()
-
-    let container: NSPersistentContainer
-
-    init(inMemory: Bool = false) {
-        container = NSPersistentContainer(name: "ClashOps")
-
-        // ✅ Enable automatic lightweight migration
-        if let description = container.persistentStoreDescriptions.first {
-            description.setOption(true as NSNumber, forKey: NSMigratePersistentStoresAutomaticallyOption)
-            description.setOption(true as NSNumber, forKey: NSInferMappingModelAutomaticallyOption)
-            if inMemory {
-                description.url = URL(fileURLWithPath: "/dev/null")
-            }
-        }
-
-        container.loadPersistentStores { _, error in
-            if let error = error as NSError? {
-                fatalError("Unresolved error \(error), \(error.userInfo)")
-            }
-        }
-
-        // ✅ Auto-merge context changes
-        container.viewContext.automaticallyMergesChangesFromParent = true
-        container.viewContext.mergePolicy = NSMergeByPropertyObjectTrumpMergePolicy
-    }
-
-}
-
